pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "climasense"
version = "0.1.0"
description = "ClimaSense: Agentic Climate Risk Intelligence for Smallholder Farmers — Gemma 4 Good Hackathon"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.45.0",
    "vllm>=0.7.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "ipykernel", "jupyter"]
kaggle = ["kaggle-benchmarks"]

[tool.setuptools.packages.find]
where = ["src"]
